Do not use sudo cookbook
This was used to allow passwordless sudo. It seems like it was never
working. The provider does not actually copy the /etc/sudoers file
itself, only creates files in /etc/sudoers.d, which were not getting
executed. The provider also does not work on SmartOS because sudoers is
in /opt.